// frontend/src/lib/api/schema.ts
export interface components {
    schemas: {
// ...
        MonitorRead: {
            /** Name */
            name: string;
            /** Target */
            target: string;
            /**
             * Check Type
             * @default http
             */
            check_type: string;
            /**
             * Check Config
             * @default {}
             */
            check_config: {
                [key: string]: unknown;
            };
            /** Frequency */
            frequency: number;
            /** Id */
            id: number;
            /** State */
            state: string;
            /** Last State */
            last_state: string | null;
            /** Last Checked At */
            last_checked_at: string | null;
            /** Consecutive Failures */
            consecutive_failures: number;
            /**
             * Created At
             * Format: date-time
             */
            created_at: string;
            /** User Id */
            user_id: number;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}